requirements.txt
python-telegram-bot>=20.0
moviepy
gTTS
langdetect
google-api-python-client
google-auth-oauthlib